requirements.txt
torch==2.11.0
transformers==5.6.2
accelerate==1.13.0
PyMuPDF==1.27.2.2
PyYAML==6.0.3
pandas==3.0.2
pytest==9.0.3